Remove dead code from hinge loss plugin
