+++ plugins/hinge_loss/schemas.py
@@ -16,7 +16,20 @@ from dataclasses import dataclass
 
 import marshmallow as ma
 
-from qhana_plugin_runner.api.util import FrontendFormBaseSchema, MaBaseSchema, FileUrl
+from qhana_plugin_runner.api.util import FileUrl, FrontendFormBaseSchema, MaBaseSchema
+
+
+@dataclass
+class CallbackUrl:
+    callback: str
+
+
+class CallbackUrlSchema(MaBaseSchema):
+    callback = ma.fields.URL(required=True, allow_none=False)
+
+    @ma.post_load
+    def make_object(self, data, **kwargs):
+        return CallbackUrl(**data)
 
 
 @dataclass
